+++ wand/cdefs/magick_image.py
@@ -487,6 +487,14 @@ def load(lib, IM_VERSION):
         lib.MagickGetImageFeatures.restype = c_void_p
     lib.MagickGetImageFilename.argtypes = [c_void_p]
     lib.MagickGetImageFilename.restype = c_void_p
+    if IM_VERSION >= 0x710:
+        try:
+            lib.MagickGetImageFilter.argtypes = [c_void_p]
+            lib.MagickGetImageFilter.restype = c_int
+        except AttributeError:
+            lib.MagickGetImageFilter = None
+    else:
+        lib.MagickGetImageFilter = None
     lib.MagickGetImageFormat.argtypes = [c_void_p]
     lib.MagickGetImageFormat.restype = c_void_p
     lib.MagickGetImageFuzz.argtypes = [c_void_p]
@@ -984,6 +992,14 @@ def load(lib, IM_VERSION):
     lib.MagickSetImageExtent.restype = c_bool
     lib.MagickSetImageFilename.argtypes = [c_void_p, c_char_p]
     lib.MagickSetImageFilename.restype = c_bool
+    if IM_VERSION >= 0x710:
+        try:
+            lib.MagickSetImageFilter.argtypes = [c_void_p, c_int]
+            lib.MagickSetImageFilter.restype = c_bool
+        except AttributeError:
+            lib.MagickSetImageFilter = None
+    else:
+        lib.MagickSetImageFilter = None
     lib.MagickSetImageFormat.argtypes = [c_void_p, c_char_p]
     lib.MagickSetImageFormat.restype = c_bool
     lib.MagickSetImageFuzz.argtypes = [c_void_p, c_double]